Adjusting installer scripts to utilize git IzPack repository
git-svn-id: file:///home/svn/repos/tigase-server/trunk@2458 7d282ba1-3ae6-0310-8f9b-c9008a0864d2

+++ scripts/generate-installer.sh
@@ -3,6 +3,11 @@
 IZPACK_DIR="installer/izpack.patched"
 #IZPACK_DIR="/Applications/IzPack"
 
+# create packages directory
+if [ ! -e packages ] ; then
+	mkdir packages || exit -1
+fi
+
 export TIGVER=`grep -m 1 "Tigase-Version:" MANIFEST.MF | sed -e "s/Tigase-Version: \(.*\)/\\1/"`
 sed -e "s/<appversion>\([^<]*\)<\/appversion>/<appversion>$TIGVER<\/appversion>/" \
     src/main/izpack/install.xml > src/main/izpack/install_copy.xml
